[project @ 2003-06-09 15:37:37 by simonpj] ------------------------- 	Fix the crossDllArg crash 	-------------------------
Test codeGen/should_compile/cg004 tests for this one.  The problem was
that the test for static-ness (i.e. no need to CAFify the thing) encountered
a form we didn't previously expect.  (See cg004)

This fix tidies up CoreUtils.rhsIsNonUpd, which is the original entry point,
renaming it CoreUtils.hasNoRedexes