[project @ 2003-06-12 16:50:19 by simonpj] ---------------------------------- 	Fix the no-GHC.Err problem in GHCi 	----------------------------------
Merge this to the stable branch.

Pattern-match failure error-ids are wired-in, which means the
GHC.Err interface isn't loaded, which in turn confused the linker
when tried to find what package GHC.Err is from.

Now we exclude wired-in names from the free variables looked at
by the linker (in Linker.linkExpr), and make sure the base package
is loaded unconditionally (DriverState.initPackageList).