[project @ 2003-07-21 14:33:19 by simonmar] :cd now causes all modules to be unloaded.
This is the easy fix for a problem whereby changing directories will
cause GHC to try to demand-load objects from the wrong place, because
it is using relative pathnames.

It's not obvious what the "right" thing to do is if the user changes
directory with :cd and then does a :reload.  Perhaps the right thing
to do is to make the search path "sticky", so that it stays relative
to the original current directory.  This would probably require some
re-engineering of GHC to only use absolute paths internally.