[project @ 2003-08-15 11:31:02 by simonmar] Fix a cause of non-linearity when compiling code with large amounts of static data, such as long lists of integers.
The problem was that CorePrep was calling allLazy on the list of
floated bindings at each level of a nested let, and allLazy runs over
all the bindings floated so far, hence O(n^2) performance.  The fix is
to cache the result of allLazy along with the list of floated
bindings.