[project @ 2003-08-20 10:22:44 by simonmar] Use the wide-char classifications from the C library if available. This gives us Unicode-aware isLower, isUpper, isAlpha etc.
On Unix, you have to set your locale to something.  This is usually
done by setting the environment variable LANG, eg.

  export LANG=en

This stuff *should* also work on Windows, except that Windows uses a
16-bit wchar_t so will get it wrong for characters > '\xffff'.  However,
I figured it was better to use the system-supplied functionality
rather than trying to implement this stuff ourselves.