[project @ 2003-09-23 15:15:02 by simonpj] --------------------------      Move demand-zapping code to where it belongs 	   --------------------------
A rather subtle point in the simplifier concerns the zapping of demand-info
when the RHS of a binding is a value.  This used to be tucked away inside
IdInfo where it was hard to find.  This commit moves the code to Simplify,
so it occurs where you'd look for it.  Along with copious comments.

See the zapDemandInfo in Simplify.completeLazyBind