[project @ 2003-10-08 09:42:34 by wolfgang] Mac OS X/PowerPC: Learn to cope with out-of-range relative jumps.
PowerPC relative branch instructions have a 24 bit displacement field.
As PPC code is always 4-byte-aligned, this yields a +-32MB range.
If a particular imported symbol is outside this range, we have to redirect
the jump to a short piece of new code that just loads the 32bit absolute
address and jumps there.