[project @ 2004-01-28 10:04:25 by simonpj] ---------------------------------------   	 Fix the WinME/98/95 double-backslash bug 	  ---------------------------------------
		Merge to stable branch, at least
		once we are sure it's right

Backslashes should not be escaped in command-line arguments for rawSystem,
on Win32.  This only actually causes failures on WinME/98/95, and we can't
test that since we don't have it.  But this fix seems right regardless.

There are extensive new comments in libraries/base/System/Cmd.hs which
describe the issues.