[project @ 2004-02-24 10:34:21 by simonmar] Fix line-comments.  The problem was that eg.
   --->

at the beginning of a line was matched by the line-comment rule,
because the varsym rule isn't valid at this point (we're in the bol
state).  Fix is to split the line-comment regex into two:

"--"\-* [^$symbol] .*			;
"--"\-* / { atEOL }			;

so a sequence of dashes on its own is only treated as a comment if
we're at the end of the line (or file).  Otherwise the dashes must be
followed by a non-symbol character.