[project @ 2004-02-24 15:57:52 by simonpj] ---------------------------------------   	 Record dependency on Template Haskell package 	  ---------------------------------------
An unforseen consequence of making the Template Haskell package separate
is that we need to record dependency on the package, even if no TH module
is imported.  So we carry round (another) mutable variable tcg_th_used in
the tyepchecker monad, and zap it when $(...) and [| ... |] are used.

I did a little tidy-up and documentation in ListSetOps too