[project @ 2004-03-11 14:34:22 by simonpj] Fix a nasty and long-standing bug in the handling of functional dependencies.
The story is told in comments with TcSimplify.tcSimplifyRestricted.  We were
simpifying a group of constraints *twice*: once to establish the type vars to
quantify over, and once "for real" but less brutally.  Unfortunately, the
less-brutally part meant that we did less improvement, which in turn meant
that an invariant (no captured constraints) was violated.  Consequential
bizarre results.

The test is typecheck/should_compile/tc177