[project @ 2004-04-21 12:45:54 by simonpj] Do a much better job of slurping RULES.
Now that stuff is slurped in lazily, as the simplifier pokes on it,
we may not get the rules as early as we might wish.  In the current
HEAD, no new rules are slurped in after the beginning of SimplCore,
and that means we permanently miss many rules.

This commit arranges that every time round the simplifier loop we
slurp in any new rules, and put them into the in-scope set, where the
simplifier can find them.

It's still possible that a rule might be slurped in a little later than
in earlier versions of GHC, leading to more simplifier iterations,
but let's see if that turns out to be a problem in practice.