[project @ 2004-06-02 08:25:10 by simonpj] -----------------------------------------------        Record whether data constructors are declared infix 	-----------------------------------------------
This allows us to generate the InfixC form in Template Hasekll.
And for 'deriving' Read and Show, we now read and parse the infix
form iff the constructor was declared infix, rather than just if
it does not have the default fixity (as before).

IfaceSyn changes slightly, so that IfaceConDecl can record their
fixity, so there are trivial changes scattered about, and
you'll need to recompile everything.

In TysWiredIn I took the opportunity to simplify pcDataCon slightly,
by eliminating the unused Theta argument.